Make handling of Approx.epsilon more sane
Comparing floating point numbers approximately:

We have numeric_limits<T>::epsilon() which is the difference between 1.0
and the next representable number as T (~1.19e-7 for floats). The next
representable number depends on the floating point exponent. For some
arbitrary value the next available number is approximately
(epsilon*value). So (value + epsilon*value) is approximately the next
available number.

When comparing floating point numbers, it makes sense to compare against
how far apart adjacent values are. The numbers can said to be
approximately correct if they are within a few representable values.
Comparing x to y as: (abs(x-y) < scale*epsilon*y) is basically saying
x==y if x is within scale number of representable values of y.

The old calculation is (abs(x-y) < epsilon*100*(scale+y)) which says
x==y if x is within 100 representable values of y + epsilon*100*scale.
Here scale is 1.0 by default, so the comparison depends on 100
representable values of 1.0.

Comparing the two methods by example (note the new method has
scale=100.0 by default):
1.0 == Approx (1.001)
Old:
    0.001 < 1.19e-7 * 100 * (1.0+1.001)
    0.001 < 1.19e-5 * (2.001)
    0.001 < 2.381-5
New:
    0.001 < 1.19e-7 * 100 * 1.001
    0.001 < 1.19e-5 * (1.001)
    0.001 < 1.191-5

1.0e-6 == Approx (1.001e-6)
Old:
    0.001e-6 < 1.19e-7 * 100 * (1.0+1.001e-6)
    0.001e-6 < 1.19e-5 * (1.000001001)
    0.001e-6 < 1.19e-5 # Note the discrepancy here!
New:
    0.001e-6 < 1.19e-7 * 100 * (1.001e-6)
    0.001e-6 < 1.19e-5 * (1.001e-6)
    0.001e-6 < 1.191e-11

1.0e6 == Approx (1.001e6)
Old:
    0.001e6 < 1.19e-7 * 100 * (1.0+1.001e6)
    0.001e6 < 1.19e-5 * (1.001001e6)
    0.001e6 < 11.91
New:
    0.001e6 < 1.19e-7 * 100 * (1.001e6)
    0.001e6 < 1.19e-5 * (1.001e6)
    0.001e6 < 11.91

Note that in every case the numbers aren't relatively close to each
other. For normal/large numbers the bound for comparison is quite
similar for each version and the result is that the numbers are not
equal. But for the small numbers, the old calculation has a lower limit
of 1.19e-5 and ends up calling the numbers equal even though they are
more than 8000 representable values apart.

P.S. As written above the new method will fail all comparisons against
zero unless the other number is exactly zero (because epsilon*scale*zero
is zero). The old method does not have that problem. This can be worked
around by using the margin. However, in the code the right hand side of
the comparison uses max(x,y) instead of just y. This alleviates the
issue of comparing against zero, but may not be the best default.

Of course, the default values can be changed by the user, but not the
calculation.

P.P.S The entire operation should probably not cast everything as
doubles, and use the float epsilon. Apples to apples.

P.P.P.S This will make some of the current self tests fail. They should
be updataed, and more should be added for small/large number
comparisons.